+++ app/src/main/res/drawable/shape_rect8_elevation_level01.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android"
+    android:shape="rectangle">
+
+    <solid android:color="@color/elevation_level01" />
+    <corners android:radius="4dp" />
+</shape>
