+++ app/src/main/res/layout/activity_access_location.xml
@@ -1,0 +1,90 @@
+<?xml version="1.0" encoding="utf-8"?>
+<layout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto">
+
+    <data>
+    </data>
+
+    <androidx.constraintlayout.widget.ConstraintLayout
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:background="@color/elevation_level02"
+        android:theme="@style/Theme.TeumTeum"
+        >
+
+        <androidx.constraintlayout.widget.ConstraintLayout
+            android:layout_width="308dp"
+            android:layout_height="wrap_content"
+            android:paddingTop="40dp"
+            android:paddingBottom="20dp"
+            android:background="@drawable/shape_rect8_elevation_level01"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintTop_toTopOf="parent"
+            app:layout_constraintBottom_toBottomOf="parent"
+            >
+
+            <TextView
+                android:id="@+id/tv_title"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:text="@string/onboarding_tv_access_location_title"
+                android:textAppearance="@style/ta.headline.6"
+                android:textColor="?attr/color_text_headline_primary"
+                app:layout_constraintStart_toStartOf="parent"
+                app:layout_constraintEnd_toEndOf="parent"
+                app:layout_constraintTop_toTopOf="parent"
+                />
+
+            <TextView
+                android:id="@+id/tv_subtitle"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:text="@string/onboarding_tv_access_location_subtitle"
+                android:textAppearance="@style/ta.body.2"
+                android:layout_marginTop="8dp"
+                android:textColor="?attr/color_text_body_teritary"
+                app:layout_constraintStart_toStartOf="parent"
+                app:layout_constraintEnd_toEndOf="parent"
+                app:layout_constraintTop_toBottomOf="@id/tv_title"
+                />
+
+            <TextView
+                android:id="@+id/btn_decline"
+                android:layout_width="0dp"
+                android:layout_height="56dp"
+                android:gravity="center"
+                android:textAppearance="@style/ta.headline.6"
+                android:textColor="?attr/color_text_button_alternative"
+                android:text="@string/onboarding_tv_access_location_decline"
+                android:background="@drawable/shape_rect4_button_alternative"
+                app:layout_constraintStart_toStartOf="parent"
+                app:layout_constraintEnd_toStartOf="@id/btn_allow"
+                android:layout_marginEnd="4dp"
+                android:layout_marginStart="20dp"
+                app:layout_constraintTop_toBottomOf="@id/tv_subtitle"
+                app:layout_constraintBottom_toBottomOf="parent"
+                android:layout_marginTop="28dp"
+                />
+
+            <TextView
+                android:id="@+id/btn_allow"
+                android:layout_width="0dp"
+                android:layout_height="56dp"
+                android:gravity="center"
+                android:textAppearance="@style/ta.headline.6"
+                android:textColor="?attr/color_text_button_primary_default"
+                android:text="@string/onboarding_tv_access_location_allow"
+                android:background="@drawable/shape_rect4_button_active"
+                app:layout_constraintStart_toEndOf="@id/btn_decline"
+                android:layout_marginStart="4dp"
+                android:layout_marginEnd="20dp"
+                app:layout_constraintEnd_toEndOf="parent"
+                app:layout_constraintTop_toBottomOf="@id/tv_subtitle"
+                app:layout_constraintBottom_toBottomOf="parent"
+                android:layout_marginTop="28dp"
+                />
+
+        </androidx.constraintlayout.widget.ConstraintLayout>
+    </androidx.constraintlayout.widget.ConstraintLayout>
+</layout>
