+++ app/src/main/res/values/strings.xml
@@ -25,6 +25,11 @@
     <string name="onboarding_tv_networking_subtitle">회사/학교 인증, 지인 인증, 리뷰 등으로\n안전한 네트워킹을 즐겨보세요</string>
     <string name="shake_title">세 분의 관심사의 틈을 채워\n대화 주제를 조합하는 중이에요</string>
 
+    <string name="onboarding_tv_access_location_title">위치 권한을 허용해주세요</string>
+    <string name="onboarding_tv_access_location_subtitle">지역 기반 기능 사용 시 위치 권한이 필요합니다.</string>
+    <string name="onboarding_tv_access_location_allow">허용</string>
+    <string name="onboarding_tv_access_location_decline">허용 안함</string>
+
     <string name="signin_tv_kakao">카카오톡으로 시작하기</string>
     <string name="signin_tv_naver">네이버로 시작하기</string>
